% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_emission_profile_financial.R
\name{bar_plot_emission_profile_financial}
\alias{bar_plot_emission_profile_financial}
\title{Create a horizontal stacked bar chart with financial data}
\usage{
bar_plot_emission_profile_financial(
  data,
  benchmarks = c("all", "isic_4digit", "tilt_sector", "unit", "unit_isic_4digit",
    "unit_tilt_sector"),
  mode = c("equal_weight", "worst_case", "best_case")
)
}
\arguments{
\item{data}{A data frame like \link{financial}.}

\item{benchmarks}{A character vector specifying the benchmarks for which the
emission profiles will be plotted. The user can choose from one to several
benchmark(s) to be plotted.}

\item{mode}{The mode of financial data to plot.
It can be one of "equal_weight", "worst_case" or "best_case".
If nothing is chosen, "equal_weight" is the default case.}
}
\value{
A \link{ggplot} object.
}
\description{
Generate a horizontal stacked bar chart showing the distribution of
the emission risk profiles risks for one or several benchmarks.
}
\examples{
benchmarks <- c("all", "unit", "isic_4digit")
bar_plot_emission_profile_financial(financial, benchmarks, "equal_weight")
}
