% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_xctr.R
\name{bar_plot_xctr}
\alias{bar_plot_xctr}
\title{Create a horizontal stacked bar chart without financial data}
\usage{
bar_plot_xctr(
  data,
  benchmarks = c("all", "unit", "tilt_sec", "unit_tilt_sec", "isic_sec", "unit_isic_sec")
)
}
\arguments{
\item{data}{A data frame like \link{without_financial}}
}
\value{
A \link{ggplot} object.
}
\description{
Generate a horizontal stacked bar chart showing the distribution of
the emission risk profiles risks for one or several benchmarks
}
\examples{
benchmarks <- c("all", "unit", "isic_sec")
bar_plot_xctr(without_financial, benchmarks)
}
