% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{benchmarks}
\alias{benchmarks}
\title{Benchmarks used in the functions' arguments.}
\usage{
benchmarks()
}
\description{
Benchmarks used in the functions' arguments.
}
\examples{
benchmarks()
}
\keyword{internal}
