% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check crucial columns in data}
\usage{
check_data(data, crucial)
}
\arguments{
\item{data}{A data frame to be checked}

\item{crucial}{A character vector of column names that are crucial}
}
\description{
Checks for the presence of crucial columns in the data.
}
