% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_financial}
\alias{example_financial}
\title{Construct example financial data}
\usage{
example_financial(...)
}
\arguments{
\item{...}{Passed to \code{\link[=tibble]{tibble()}}.}
}
\value{
A \code{\link[=tibble]{tibble()}}
}
\description{
Construct example financial data
}
\examples{
example_financial()

# Add rows
example_financial(bank_id = 1:2)

# Add columns
example_financial(new = 1)

# Remove columns
example_financial(bank_id = NULL)

# Pass column names stored in vectors
id <- "bank_id"
example_financial(!!id := "c")
}
\seealso{
Other example datasets: 
\code{\link{example_without_financial}()}
}
\concept{example datasets}
\keyword{internal}
