% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\name{example_without_financial}
\alias{example_without_financial}
\title{Construct example without financial data}
\usage{
example_without_financial(...)
}
\arguments{
\item{...}{Passed to \code{\link[=tibble]{tibble()}}.}
}
\value{
A \code{\link[=tibble]{tibble()}}
}
\description{
Construct example without financial data
}
\examples{
example_without_financial()

#' # Add columns
example_without_financial(new = 1)

#' # Pass column names stored in vectors
company_name <- "company_name"
example_financial(!!company_name := "c")
}
\seealso{
Other example datasets: 
\code{\link{example_financial}()}
}
\concept{example datasets}
\keyword{internal}
