% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial.R
\docType{data}
\name{financial}
\alias{financial}
\title{Financial toy data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 114 rows and 13 columns.
}
\source{
\verb{@lindadelacombaz}
}
\usage{
financial
}
\description{
Financial toy data
}
\examples{
financial
}
\keyword{datasets}
