% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financial_toy_data.R
\docType{data}
\name{financial_toy_data}
\alias{financial_toy_data}
\title{Sankey toy data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 19 rows and 12 columns.
}
\source{
\verb{@lindadelacombaz}
}
\usage{
financial_toy_data
}
\description{
Sankey toy data
}
\examples{
financial_toy_data
}
\keyword{datasets}
