% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/german_map.R
\name{german_map}
\alias{german_map}
\title{Create a German map with the risk color of each region}
\usage{
german_map(
  data,
  benchmark = c("all", "unit", "tilt_sec", "unit_tilt_sec", "isic_sec", "unit_isic_sec"),
  finance_weight = c("equal_weight", "worst_case", "best_case")
)
}
\arguments{
\item{data}{A data frame like \link{financial}}

\item{benchmark}{The mode of benchmark to plot.
It can be one of "all", "unit" or "tilt_sec", "unit_tilt_sec", "isic_sec"
or "unit_isic_sec". If nothing is chosen, "all" is the default mode.}

\item{finance_weight}{The mode of financial data to plot.
It can be one of "equal_weight", "worst_case" or "best_case". If nothing is
chosen, "equal_weight" is the default mode.}
}
\value{
A ggplot2 object representing the german data plot.
}
\description{
Create a German map with the risk color of each region
}
\examples{
# Plot a German map with a "unit" benchmark and equal_weight finance
german_map(financial, benchmark = "unit")
}
