% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{grouping_emission}
\alias{grouping_emission}
\title{grouping_emission used in the functions' arguments.}
\usage{
grouping_emission()
}
\description{
grouping_emission used in the functions' arguments.
}
\examples{
grouping_emission()
}
\keyword{internal}
