% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_region_risk_financial.R
\name{map_region_risk_financial}
\alias{map_region_risk_financial}
\title{Create a map with the risk color of each region (NUTS3 granularity), with
financial data}
\usage{
map_region_risk_financial(
  data,
  country_code = c("DE"),
  benchmark = c("all", "unit", "tilt_sec", "unit_tilt_sec", "isic_sec", "unit_isic_sec"),
  finance_weight = c("equal_weight", "worst_case", "best_case")
)
}
\arguments{
\item{data}{A data frame like \link{financial}}

\item{country_code}{Country code (ISO 3166 alpha-2) for which the map will be
plotted.}

\item{benchmark}{The mode of benchmark to plot.
It can be one of "all", "unit" or "tilt_sec", "unit_tilt_sec", "isic_sec"
or "unit_isic_sec". If nothing is chosen, "all" is the default mode.}

\item{finance_weight}{The mode of financial data to plot (#TODO : fix financial columns).
It can be one of "equal_weight", "worst_case" or "best_case". If nothing is
chosen, "equal_weight" is the default mode.}
}
\value{
A ggplot2 object representing the country data plot.
}
\description{
Create a map with the risk color of each region (NUTS3 granularity), with
financial data
}
\examples{
# Plot a German with a "unit" benchmark and equal_weight finance
map_region_risk_financial(financial, country_code = "DE", benchmark = "unit")
}
