% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nuts_de.R
\docType{data}
\name{nuts_de}
\alias{nuts_de}
\title{Correspondence table between German postcodes and NUTS level 3 codes}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 8320 rows and 2 columns.
}
\source{
"© European Union - GISCO, 2021, postal code point dataset,
Licence CC-BY-SA 4.0 available at https://ec.europa.eu/eurostat/web/gisco/geodata/reference-data."
}
\usage{
nuts_de
}
\description{
Correspondence table between German postcodes and NUTS level 3 codes
}
\examples{
nuts_de
}
\keyword{internal}
