% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey_old.R
\name{plot_sankey_old}
\alias{plot_sankey_old}
\title{Create a sankey plot}
\usage{
plot_sankey_old(data, with_company = TRUE)
}
\arguments{
\item{data}{A data frame like \link{toy_data}.}

\item{with_company}{Logical. If TRUE, will plot a node with the company name.
If FALSE, will plot without the company name node.}
}
\value{
A sankey plot of class \link{sankeyNetwork}.
}
\description{
Create a sankey plot
}
\examples{
plot_sankey(toy_data)
}
