% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xctr.R
\name{plot_xctr}
\alias{plot_xctr}
\title{Plot xctr bar plots}
\usage{
plot_xctr(data, company_name = NULL, portfolio_level = FALSE)
}
\arguments{
\item{data}{A data frame like \link{xctr_toy_data}.}

\item{company_name}{A string. Name of one company in the data set, if the user
wants to have xctr plot on a company level.}

\item{portfolio_level}{A boolean. TRUE will plot on a portfolio level, FALSE
will plot on a company level, for which a company name must be provided.}
}
\value{
A \link{ggplot} object.
}
\description{
Plot xctr bar plots
}
\examples{
# Company level
plot_xctr(xctr_toy_data, "company_a")

# Portfolio level
plot_xctr(xctr_toy_data, portfolio_level = TRUE)
}
