% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xctr_company.R
\name{plot_xctr_company}
\alias{plot_xctr_company}
\title{Plot xctr data on a company level, without financial data}
\usage{
plot_xctr_company(data, company_name)
}
\arguments{
\item{data}{A data frame like \link{without_financial}.}

\item{company_name}{A string. Name of one company in the data set.}
}
\value{
A \link{ggplot} object.
}
\description{
Plot xctr data on a company level, without financial data
}
\examples{
plot_xctr_company(without_financial, "peter")
}
