% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xctr_company_financial.R
\name{plot_xctr_company_financial}
\alias{plot_xctr_company_financial}
\title{Plot xctr data on a company level with financial data}
\usage{
plot_xctr_company_financial(
  data,
  company_name,
  mode = c("equal_weight", "worst_case", "best_case", "main_activity")
)
}
\arguments{
\item{data}{A data frame like \link{financial}.}

\item{company_name}{A string. Name of one company in the data set.}

\item{mode}{A string. Name of the financial weight.}
}
\value{
A \link{ggplot} object.
}
\description{
Plot xctr data on a company level with financial data
}
\examples{
plot_xctr_company_financial(financial, company_name = "peter", mode = "best_case")
}
