% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xctr_company_level.R
\name{plot_xctr_company_level}
\alias{plot_xctr_company_level}
\title{Plot xctr data on a company level}
\usage{
plot_xctr_company_level(data, company_name)
}
\arguments{
\item{data}{A data frame like \link{pctr_toy_data} or \link{ictr_toy_data}.}

\item{company_name}{A string. Name of one company in the data set.}
}
\value{
A \link{ggplot} object.
}
\description{
Plot xctr data on a company level
}
\examples{
subset_data <- pctr_toy_data |>
 filter(company_name == sample(company_name, 1))]
plot_xctr_company_level(subset_data, company_name)
}
