% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xctr_financial.R
\name{plot_xctr_financial}
\alias{plot_xctr_financial}
\title{Plot xctr data on a company or portfolio level, with financial data}
\usage{
plot_xctr_financial(
  data,
  company_name = NULL,
  mode = c("equal_weight", "worst_case", "best_case", "main_activity")
)
}
\arguments{
\item{data}{A data frame like \link{financial}}

\item{company_name}{(optional) The name of the specific company to plot the
financial data for. If NULL, the function will plot the financial data for the portfolio.}

\item{mode}{The mode of financial data to plot.
It can be one of "equal_weight", "worst_case", "best_case", or "main_activity".}
}
\value{
A ggplot2 object representing the financial data plot.
}
\description{
Plots the financial data for either a specific company or for the portfolio as a whole.
}
\examples{
# Example 1: Plot financial data for a specific company
plot_xctr_financial(data = financial, company_name = "peter", mode = "equal_weight")

# Example 2: Plot portfolio-level financial data
plot_xctr_financial(data = financial, mode = "worst_case")

}
