% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xctr_impl.R
\name{plot_xctr_impl}
\alias{plot_xctr_impl}
\title{Generate financial data plot}
\usage{
plot_xctr_impl(data, y_label, benchmark_var)
}
\arguments{
\item{data}{A data frame with prepared financial data}

\item{y_label}{The label for the y-axis}

\item{benchmark_var}{The variable for facet wrapping}
}
\value{
A ggplot2 object representing the financial data plot
}
\description{
Generates a ggplot2 object representing the financial data plot.
}
