% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xctr_portfolio.R
\name{plot_xctr_portfolio}
\alias{plot_xctr_portfolio}
\title{Plot xctr data on a portfolio level}
\usage{
plot_xctr_portfolio(data)
}
\arguments{
\item{data}{A data frame like \link{without_financial}.}
}
\value{
A \link{ggplot} object.
}
\description{
Plot xctr data on a portfolio level
}
\examples{
plot_xctr_portfolio(without_financial)
}
