% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xctr_portfolio_financial.R
\name{plot_xctr_portfolio_financial}
\alias{plot_xctr_portfolio_financial}
\title{Plot xctr data on a portfolio level, with financial data}
\usage{
plot_xctr_portfolio_financial(
  data,
  mode = c("equal_weight", "worst_case", "best_case", "main_activity")
)
}
\arguments{
\item{data}{A data frame like \link{financial}.}

\item{mode}{A string. Name of the financial weight.}
}
\value{
A \link{ggplot} object.
}
\description{
Plot xctr data on a portfolio level, with financial data
}
\examples{
plot_xctr_portfolio_financial(financial, mode = "worst_case")
}
