% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xctr_portfolio.R
\name{plot_xctr_portfolio_level}
\alias{plot_xctr_portfolio_level}
\title{Plot xctr data on a company level}
\usage{
plot_xctr_portfolio_level(data)
}
\arguments{
\item{data}{A data frame like \link{xctr_toy_data}.}
}
\value{
A \link{ggplot} object.
}
\description{
Plot xctr data on a company level
}
\examples{
plot_xctr_com(xctr_toy_data)
}
