% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_financial_data.R
\name{prepare_financial_data}
\alias{prepare_financial_data}
\title{Prepare financial data for plotting}
\usage{
prepare_financial_data(data)
}
\arguments{
\item{data}{A data frame like \link{financial}}
}
\value{
A processed data frame ready for plotting
}
\description{
Prepares the financial data for plotting by handling missing values and transforming risk categories.
}
