% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot_financial.R
\name{scatter_plot_financial}
\alias{scatter_plot_financial}
\title{Create a scatter plot with financial data}
\usage{
scatter_plot_financial(
  data,
  benchmarks = benchmarks(),
  mode = c("equal_weight", "worst_case", "best_case"),
  scenario = c("IPR", "WEO"),
  year = c(2030, 2050)
)
}
\arguments{
\item{data}{A data frame like \link{financial}.}

\item{benchmarks}{A character vector specifying the benchmarks for which the
emission profiles will be plotted. The user can choose from one to several
benchmark(s) to be plotted.}

\item{mode}{The mode of financial data to plot.
It can be one of "equal_weight", "worst_case" or "best_case".
If nothing is chosen, "equal_weight" is the default case.}

\item{scenario}{A character vector specifying the scenario to be plot. It can
either be "IPR" or "WEO".}

\item{year}{A numerical specifying the year of the scenario to be plot. It
can be either 2030 or 2050.}
}
\value{
A \link{ggplot} object.
}
\description{
Generate a scatter plot comparing the emission profile and transition risks
for each bank in the portfolio.
}
\examples{
scatter_plot_financial(financial, benchmarks = c("all", "tilt_sector"),
mode = "equal_weight",
scenario = "IPR",
year = 2030)
}
