% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{switch_mode}
\alias{switch_mode}
\title{Switch mode function}
\usage{
switch_mode(mode)
}
\arguments{
\item{mode}{A character string specifying the mode to be switched.}
}
\value{
A character string representing the finance-related mode name that corresponds
to the input mode. If the input mode is not one of the supported options, NULL is returned.
}
\description{
Given a mode, this function returns the corresponding finance-related mode name
in the data set.
}
\examples{
switch_mode("equal_weight")
# Returns: "equal_weight_finance"

switch_mode("worst_case")
# Returns: "worst_case_finance"
}
