% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltPlot-package.R
\docType{package}
\name{tiltPlot-package}
\alias{tiltPlot}
\alias{tiltPlot-package}
\title{tiltPlot: Plots for the 'TILT' Project}
\description{
Plots for the 'TILT' project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://2degreesinvesting.github.io/tiltPlot/}
  \item \url{https://github.com/2DegreesInvesting/tiltPlot}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/tiltPlot/issues}
}

}
\author{
\strong{Maintainer}: Kalash Singhal \email{kalash@2degrees-investing.org}

Authors:
\itemize{
  \item Linda Delacombaz \email{linda@2degrees-investing.org}
  \item Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})
  \item Anne Schoenauer \email{anne.schoenauer@2degrees-investing.org} (\href{https://orcid.org/0000-0002-4576-8799}{ORCID})
  \item Tilman Trompke \email{tilman@2degrees-investing.org} (\href{https://orcid.org/0009-0003-3351-5131}{ORCID})
}

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
