% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltplot_theme.R
\name{tiltplot_theme}
\alias{tiltplot_theme}
\title{TiltPlot theme for the graphs}
\usage{
tiltplot_theme()
}
\value{
An object of class "theme", "gg".
}
\description{
TiltPlot theme for the graphs
}
\examples{
library(ggplot2, warn.conflicts = FALSE)

ggplot(mtcars) +
  geom_histogram(aes(mpg), bins = 10) +
  tiltplot_theme()
}
