% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tweak_xctr_plot.R
\name{tweak_xctr_plot}
\alias{tweak_xctr_plot}
\title{Apply customizations to a ggplot2 plot}
\usage{
tweak_xctr_plot(plot)
}
\arguments{
\item{plot}{A ggplot2 plot object}
}
\value{
A modified ggplot2 plot object
}
\description{
Adds specific geoms and themes to the provided plot.
}
