% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/without_financial.R
\docType{data}
\name{without_financial}
\alias{without_financial}
\title{Without financial data set}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 114 rows and 12 columns.
}
\source{
\verb{@lindadelacombaz}
}
\usage{
without_financial
}
\description{
Without financial data set
}
\examples{
without_financial
}
\keyword{datasets}
