% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xctr_toy_data.R
\docType{data}
\name{xctr_toy_data}
\alias{xctr_toy_data}
\title{XCTR toy data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6 rows and 4 columns.
}
\source{
\verb{@lindadelacombaz}
}
\usage{
xctr_toy_data
}
\description{
XCTR toy data
}
\examples{
xctr_toy_data
}
\keyword{datasets}
