% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_emission_profile.R
\name{bar_plot_emission_profile}
\alias{bar_plot_emission_profile}
\title{Create a horizontal stacked bar chart without financial data}
\usage{
bar_plot_emission_profile(
  data,
  grouping_emission = grouping_emission(),
  mode = modes(),
  scenario = scenarios(),
  year = years(),
  risk_category = risk_category()
)
}
\arguments{
\item{data}{A data frame like \link{without_financial}.}

\item{grouping_emission}{A character vector specifying the benchmarks for which the
emission profiles will be plotted. The user can choose from one to several
benchmark(s) to be plotted.}

\item{mode}{A character vector: equal_weight, best_case, worst_case.}

\item{scenario}{A character vector: 1.5C RPS, NZ 2050.}

\item{year}{A character vector: 2030, 2050.}

\item{risk_category}{A character vector.}
}
\value{
A \link{ggplot} object.
}
\description{
Generate a horizontal stacked bar chart showing the distribution of
the emission risk profiles risks for one or several benchmarks.
}
\examples{
grouping_emission <- c("unit")
bar_plot_emission_profile(without_financial, grouping_emission, risk_category = "emission_category")
}
