% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bar_plot_emission_profile.R
\name{bar_plot_emission_profile}
\alias{bar_plot_emission_profile}
\title{Create a horizontal stacked bar chart without financial data}
\usage{
bar_plot_emission_profile(data, benchmarks = benchmarks(), mode = modes())
}
\arguments{
\item{data}{A data frame like \link{without_financial}.}

\item{benchmarks}{A character vector specifying the benchmarks for which the
emission profiles will be plotted. The user can choose from one to several
benchmark(s) to be plotted.}

\item{mode}{A character vector: equal_weight_emission_profile, best_case_emission_profile, worst_case_emission_profile.}
}
\value{
A \link{ggplot} object.
}
\description{
Generate a horizontal stacked bar chart showing the distribution of
the emission risk profiles risks for one or several benchmarks.
}
\examples{
benchmarks <- c("all", "unit", "isic_4digit")
bar_plot_emission_profile(without_financial, benchmarks)
}
