% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R, R/utils-test.R
\name{example_financial}
\alias{example_financial}
\title{Construct example financial data}
\usage{
example_financial(
  bank_id = "a",
  amount_total = 10,
  company_name = "b",
  emission_profile = "medium",
  benchmark = "c",
  equal_weight_finance = 10,
  worst_case_finance = 10,
  best_case_finance = 10,
  ...
)

example_financial(
  bank_id = "a",
  amount_total = 10,
  company_name = "b",
  emission_profile = "medium",
  benchmark = "c",
  equal_weight_finance = 10,
  worst_case_finance = 10,
  best_case_finance = 10,
  ...
)
}
\arguments{
\item{...}{Passed to \code{\link[=tibble]{tibble()}}.}
}
\value{
A \code{\link[=tibble]{tibble()}}

A \code{\link[=tibble]{tibble()}}
}
\description{
Construct example financial data

Construct example financial data
}
\examples{
example_financial()

# Add rows
example_financial(bank_id = 1:2)

# Add columns
example_financial(new = 1)

# Remove columns
example_financial(bank_id = NULL)

# Pass column names stored in vectors
id <- "bank_id"
example_financial(!!id := "c")
example_financial()
}
\concept{example datasets}
\keyword{internal}
