% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-test.R
\name{example_financial}
\alias{example_financial}
\title{Construct example financial data}
\usage{
example_financial(
  bank_id = "a",
  amount_total = 10,
  company_name = "b",
  emission_profile = "medium",
  benchmark = "c",
  equal_weight_finance = 10,
  worst_case_finance = 10,
  best_case_finance = 10,
  ...
)
}
\arguments{
\item{...}{Passed to \code{\link[=tibble]{tibble()}}.}
}
\value{
A \code{\link[=tibble]{tibble()}}
}
\description{
Construct example financial data
}
\examples{
example_financial()
}
\keyword{internal}
