% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Create a sankey plot}
\usage{
plot_sankey(data)
}
\arguments{
\item{data}{A data frame like \link{toy_data}.}
}
\value{
#FIXME
}
\description{
Create a sankey plot
}
\examples{
plot_sankey(toy_data)
}
