% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xctr.R
\name{plot_xctr}
\alias{plot_xctr}
\title{Plot xctr data without financial data}
\usage{
plot_xctr(data)
}
\arguments{
\item{data}{A data frame like \link{without_financial}}
}
\value{
A \link{ggplot} object.
}
\description{
Generate bar plots showing the distribution of xctr_risk_category percentages for each benchmark.
}
\examples{
plot_xctr(without_financial)
}
