% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xctr.R
\name{plot_xctr}
\alias{plot_xctr}
\title{Plot xctr data on a company or portfolio level, without financial data}
\usage{
plot_xctr(data, company_name = NULL)
}
\arguments{
\item{data}{A data frame like \link{without_financial}}

\item{company_name}{(Optional) A character string specifying the company name.
If provided, the function will generate the plot for the specified company only.
If NULL (default), the plot will be generated for the entire portfolio.}
}
\value{
A \link{ggplot} object.
}
\description{
Generate bar plots showing the distribution of xctr_risk_category percentages for each benchmark.
}
\examples{
# Example 1: Generate the plot for the entire portfolio
plot_xctr(data)

# Example 2: Generate the plot for a specific company
plot_xctr(data, company_name = "peter peasant")
}
