% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_xctr_portfolio_level.R
\name{plot_xctr_portfolio_level}
\alias{plot_xctr_portfolio_level}
\title{Plot the XCTR on a portfolio level}
\usage{
plot_xctr_portfolio_level(data)
}
\arguments{
\item{data}{A data frame like \link{xctr_toy_data}.}
}
\value{
A \link{ggplot} object.
}
\description{
Plot the XCTR on a portfolio level
}
\examples{
plot_xctr_portfolio_level(xctr_toy_data)
}
