% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltPlot-package.R
\docType{package}
\name{tiltPlot-package}
\alias{tiltPlot}
\alias{tiltPlot-package}
\title{tiltPlot: Plots for the 'TILT' Project}
\description{
Plots for the 'TILT' project.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://2degreesinvesting.github.io/tiltPlot/}
  \item \url{https://github.com/2DegreesInvesting/tiltPlot}
  \item Report bugs at \url{https://github.com/2DegreesInvesting/tiltPlot/issues}
}

}
\keyword{internal}
