% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tiltPlot-package.R
\docType{package}
\name{tiltPlot-package}
\alias{tiltPlot}
\alias{tiltPlot-package}
\title{tiltPlot: Plots for the 'TILT' Project}
\description{
Plots for the 'TILT' project.
}
\author{
\strong{Maintainer}: Linda Delacombaz \email{linda@2degrees-investing.org}

Authors:
\itemize{
  \item Anne Schoenauer \email{anne.schoenauer@2degrees-investing.org}
}

Other contributors:
\itemize{
  \item 2 Degrees Investing Initiative \email{contact@2degrees-investing.org} [copyright holder, funder]
}

}
\keyword{internal}
