% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toy_data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 2 rows and 5 columns.
}
\source{
\verb{@lindadelacombaz}
}
\usage{
toy_data
}
\description{
Toy data
}
\examples{
toy_data
}
\keyword{datasets}
